examples/google_adk_plugin/pyproject.toml
[project]
name = "agent-control-google-adk-plugin-example"
version = "0.1.0"
description = "Google ADK packaged plugin integration example for Agent Control"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agent-control-engine",
    "agent-control-evaluators",
    "agent-control-models",
    "agent-control-sdk",
    "google-adk>=1.0.0,<2.0.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["my_agent"]

[tool.uv.sources]
agent-control-sdk = { path = "../../sdks/python", editable = true }
agent-control-models = { path = "../../models", editable = true }
agent-control-engine = { path = "../../engine", editable = true }
agent-control-evaluators = { path = "../../evaluators/builtin", editable = true }
